Only collect locks metrics if relations are specified
