+++ postgres/tests/test_relations.py
@@ -1,7 +1,9 @@
 # (C) Datadog, Inc. 2010-present
 # All rights reserved
 # Licensed under Simplified BSD License (see LICENSE)
+import psycopg2
 import pytest
+from .common import HOST, DB_NAME, PORT
 
 RELATION_METRICS = [
     'postgresql.seq_scans',
@@ -130,3 +132,27 @@ def test_index_metrics(aggregator, integration_check, pg_instance):
 
     for name in IDX_METRICS:
         aggregator.assert_metric(name, count=1, tags=expected_tags)
+
+
+@pytest.mark.integration
+@pytest.mark.usefixtures('dd_environment')
+def test_locks_metrics(aggregator, integration_check, pg_instance):
+    pg_instance['relations'] = ['persons']
+    pg_instance['dbname'] = 'datadog_test'
+
+    check = integration_check(pg_instance)
+    with psycopg2.connect(host=HOST, dbname=DB_NAME, user="postgres", password="datad0g") as conn:
+        with conn.cursor() as cur:
+            cur.execute('LOCK persons')
+            check.check(pg_instance)
+
+    expected_tags = pg_instance['tags'] + [
+        'server:{}'.format(HOST),
+        'port:{}'.format(PORT),
+        'db:datadog_test',
+        'lock_mode:AccessExclusiveLock',
+        'lock_type:relation',
+        'table:persons',
+        'schema:public',
+    ]
+    aggregator.assert_metric('postgresql.locks', count=1, tags=expected_tags)
